redis: add capacity to spawn a redis cluster
When using SES, it can be useful to have an SQS for bounce email, so can
having a redis cluster be useful for storing/queing email to send them.

This is just like SQS optional and not limited to redis, as elasticache
could also create memcache cluster; but it became fairly common to have
redis usage coupled with email to make it more dynamic.